Improve column pruning for shared nodes
Since in our cannonical order, filters are expected to be under
projections, allow the column pruning rules to see through filter
nodes and clone them.

Example:

  U
 / \
P1  P2
|   |
F1  F2
 \ /
  J

Join J can be pruned so it only projects the columns needed by P1, P2,
F1 and F2. The rule is fired on J but P1 and P2 are the nodes being
replaced.